refactor: change distance to bigint

+++ db/migrate/20231121221457_create_planets.rb
@@ -2,7 +2,7 @@ class CreatePlanets < ActiveRecord::Migration[7.1]
   def change
     create_table :planets do |t|
       t.string :name, null: false
-      t.integer :distance_to_earth, null: false, comment: "Distance in megameters (1mm = 1,000,000km)"
+      t.bigint :distance_to_earth, null: false, comment: "Distance in megameters (1mm = 1,000,000km)"
 
       t.timestamps
     end
